+++ docs/CHANGELOG.md
@@ -3,3 +3,6 @@
 ## Changes in 1.0-Release:
 
 + Update: to **Tukui 18.21+**
++ Added: plugin Dispels Announce.
++ Added: plugin Interrupts Announce.
++ Added: plugin ScreenShots.